==> 1_Extract_Data.ipynb <==
**Import necessary libraries.**

from bs4 import BeautifulSoup
import pandas as pd
import os

`extract_data` is a function uses `soup` library to extract the data from  the given dataset.

def extract_data(soup):
    item_data = {}
    addition_data = []

    if soup.find('table', class_='driving-table'):
        item_data['Name'] = soup.find('h3').text.strip()
        try:
            item_data['Model'] = int(soup.find('h5').text.strip().split()[2])
            item_data['Price'] = int(soup.find('h5', class_='post-price').text.strip().split()[0])
        except (IndexError, ValueError):
            return None

        for additions_table in soup.find_all('table', class_='list_ads'):
            for row in additions_table.find_all('tr'):
                cells = row.find_all('td')
                if len(cells) >= 2:
                    attribute = cells[0].text.strip()
                    value = cells[1].text.strip()

                    addition_data.append(attribute)
                    item_data[attribute] = value
    else:
        return None

    return item_data

Read the dataset and use `extract_data` function.

folder_path = './data/'
cars_data = []

for filename in os.listdir(folder_path):
    if filename.endswith('.txt'):
        file_path = os.path.join(folder_path, filename)

        with open(file_path, 'r', encoding='utf-8') as file:
            html_content = file.read()

        soup = BeautifulSoup(html_content, 'html.parser')
        data = extract_data(soup)

        if data:
            cars_data.append(data)


Convert the extracted data into a `df` and convert it into a `.csv` file.

df = pd.DataFrame(cars_data)
csv_file_path = 'cars.csv'
df.to_csv(csv_file_path, index=False)
