==> 4_Flask_Service.ipynb <==
Import needed Libraries

from flask import Flask, request, jsonify
import numpy as np
import pandas as pd
from sklearn.tree import DecisionTreeRegressor

Create a new `df` using the `cleaned_cars.csv`

df = pd.read_csv('cleaned_cars.csv')

Training the *Decision Tree Model* on the whole data

X = df.drop('Price', axis=1)
y = df['Price']

tree_model = DecisionTreeRegressor(random_state=42)
tree_model.fit(X, y)

Create *Flask service* to predict the new car's price

app = Flask(__name__)

@app.route('/predict', methods=['POST'])
def predict():
    try:
        if request.headers['Content-Type'] == 'application/json':
            data = request.get_json(force=True)

            if 'features' in data and isinstance(data['features'], dict):
                features_dict = data['features']

                features_ordered = [features_dict.get(feature, np.nan) for feature in X.columns]
                features_df = pd.DataFrame([features_ordered])

                prediction = tree_model.predict(features_df)

                return jsonify({'prediction': prediction.tolist()})
            else:
                return jsonify({'error': 'Invalid or missing "features" data in the request'}), 400
        else:
            return jsonify({'error': 'Invalid content type. Use application/json'}), 415
    except Exception as e:
        print(f"An error occurred: {str(e)}")
        return jsonify({'error': f'An internal server error occurred: {str(e)}'}), 500

if __name__ == '__main__':
    app.run(port=5000)
